requirements.txt
# Flask and extensions
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Werkzeug==2.0.1
SQLAlchemy==1.4.23
click==7.1.2
itsdangerous==2.0.1
Jinja2==3.0.1

# Database
alembic==1.7.1

# Audio processing
librosa==0.9.1
pydub==0.25.1
numpy==1.21.2
soundfile==0.10.3.post1
scipy==1.7.1
numba==0.54.1
llvmlite==0.37.0

# Audio analysis dependencies
scikit-learn==0.24.2
joblib==1.0.1
audioread==2.1.9
pooch==1.5.1
resampy==0.2.2

# Development and deployment
python-dotenv==0.19.0
gunicorn==20.1.0
pytest==6.2.5
coverage==6.0.2

# Audio format support
ffmpeg-python==0.2.0
